+++ frontend/libs/platform/platform-ui/src/index.ts
@@ -1,3 +1,4 @@
 export * from './lib/org-currencies/org-currencies.page';
 export * from './lib/configurations/configurations.page';
 export * from './lib/smtp-settings/smtp-settings.page';
+export * from './lib/organizations/organizations.page';